feat: Add endpoint to get user by ID
This commit adds a new endpoint `/users/{id}` to retrieve a user by their ID. The endpoint queries the database for a user with the specified ID and returns the user if found. If the user does not exist, a 404 status code is returned along with the total number of users in the database.