feat: Refactor router prefixes and endpoints
This commit refactors the router prefixes and endpoints in the `post.py` and `user.py` files. The `post` router now has a prefix of `/posts` and the endpoints are updated accordingly. Similarly, the `user` router has a prefix of `/users` and the endpoints are modified to match the new prefix. This change improves the organization and readability of the code.