feat: Update login endpoint to use OAuth2PasswordRequestForm
This commit updates the `login` endpoint in the `auth.py` file to use the `OAuth2PasswordRequestForm` class from the `fastapi.security.oauth2` module. This change allows the endpoint to receive the username and password as form data instead of a JSON payload. The `user_cred` parameter is replaced with `user_cred: OAuth2PasswordRequestForm = Depends()` in the function signature. The username is accessed using `user_cred.username` instead of `user_cred.email`. The access token is now generated using the `Oauth2.create_access_token` method and returned in the response along with the token type.